+++ templates/admin/audit.php
@@ -57,6 +57,7 @@
         </div>
         <div class="admin-pagination" id="au-pagination"></div>
     </div>
+    <?php $footerInPanel = true; include __DIR__ . '/../footer.php'; ?>
 
     <div id="toast-container" class="toast-container position-fixed bottom-0 end-0 p-3"></div>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.8/dist/js/bootstrap.bundle.min.js" integrity="sha384-FKyoEForCGlyvwx9Hj09JcYn3nv7wiPVlz7YYwJrWVcXK/BmnVDxM+D2scQbITxI" crossorigin="anonymous"></script>